Brute force solution of Day 1
This is not a good solution. The first part is O(n^2) and the second
part is O(n^3), but the input is so small that it's fairly fast to
compute the result for this anyway.

It would probably be possible to short-circuit the outer loops by
first checking the smallest number in the input, and then skipping
anything in the `for i in &data` loop where `i + smallest_number >
2020`, and then doing the same thing for the two smallest numbers for
part 2.